requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
mysql-client==0.0.1
mysqlclient==2.0.3
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
